Post-merge reconciliation with latest main
Adapts the branch to main's API moves: Markdown dropped the compact
prop; ToolDetailBlocks requires fileReadContent; formatOwnerLabel
gains main's "you" short-circuit; the mention send flow's invite
helper carries the captured channel id; the composer props file ports
onCaptureSendContext. Lint deltas from main's stricter config are
fixed properly (own-bubble chip rules reordered below the standalone
rules, pass-through switch cases folded into default, pulse
suppressions re-anchored). Size ceilings are restored by extracting
the NIP-AB pairing API from tauri.ts and the find-target splice state
from ChannelScreen into useFindTargetEvents.
